+++ src/core/resources/config.yml
@@ -61,22 +61,17 @@ database:
   database: "database" # Database name
   debug: false # Debug mode for database
 
-# Configure channels for different log levels 
-# Each log level can be used in different plugin to define with channel to log to
-log:
-  minor: 
-    enabled: true
+# Where the bot should log messages
+log-channels:
+  - id: minor # Log id, used to reference the log channel in other parts of the bot
+    channel: "LOG_CHANNEL_ID" # Channel ID where the bot should log minor messages
+  - id: moderate
     channel: "LOG_CHANNEL_ID"
-  moderate:
-    enabled: true
+  - id: high
+    channel: "GUILD_ID:LOG_CHANNEL_ID" # You can also specify a guild ID to log messages in a different guild 
+  - id: major
     channel: "LOG_CHANNEL_ID"
-  high:
-    enabled: true
+  - id: critical
     channel: "LOG_CHANNEL_ID"
-  major:
-    enabled: true
-    channel: "LOG_CHANNEL_ID"
-  critical:
-    enabled: true
-    channel: "LOG_CHANNEL_ID"
+
